// src/lib/api/gen/types/ListPayouts.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type ListPayoutsPathParams = {
    /**
     * @type string
    */
    slug: string;
};

/**
 * @description Default Response
*/
export type ListPayouts200 = {
    /**
     * @type array
    */
    items: {
        /**
         * @minLength -9007199254740991
         * @maxLength 9007199254740991
         * @type integer
        */
        earnedCents: number;
        /**
         * @minLength -9007199254740991
         * @maxLength 9007199254740991
         * @type integer
        */
        settledCents: number;
        /**
         * @minLength -9007199254740991
         * @maxLength 9007199254740991
         * @type integer
        */
        balanceCents: number;
        /**
         * @type object
        */
        supplier: {
            /**
             * @type string
            */
            id: string;
            /**
             * @type string
            */
            name: string;
            /**
             * @type string
            */
            phone: string | null;
            /**
             * @type string
            */
            pixKey: string | null;
            /**
             * @type string
            */
            note: string | null;
            /**
             * @type boolean
            */
            active: boolean;
            /**
             * @type string
            */
            createdAt: string;
        };
    }[];
    /**
     * @type object
    */
    totals: {
        /**
         * @minLength -9007199254740991
         * @maxLength 9007199254740991
         * @type integer
        */
        earnedCents: number;
        /**
         * @minLength -9007199254740991
         * @maxLength 9007199254740991
         * @type integer
        */
        settledCents: number;
        /**
         * @minLength -9007199254740991
         * @maxLength 9007199254740991
         * @type integer
        */
        balanceCents: number;
    };
};

export type ListPayoutsQueryResponse = ListPayouts200;

export type ListPayoutsQuery = {
    Response: ListPayouts200;
    PathParams: ListPayoutsPathParams;
    Errors: any;
};